# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zip Code Finder API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://url.local",
            "description": "DEV Server"
        }
    ],
    "paths": {
        "/api/zip-codes/crawl": {
            "post": {
                "tags": [
                    "Zip Codes"
                ],
                "summary": "Crawl the zip codes and store them in the database.",
                "operationId": "crawl",
                "responses": {
                    "200": {
                        "description": "Successful."
                    }
                },
                "security": []
            }
        },
        "/api/zip-codes/{id}": {
            "get": {
                "tags": [
                    "Zip Codes"
                ],
                "summary": "Return the requested zip code.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data returned successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZipCode"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    }
                },
                "security": []
            }
        }
    },
    "components": {
        "schemas": {
            "Settlement": {
                "properties": {
                    "key": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "zone_type": {
                        "type": "string"
                    },
                    "settlement_type": {
                        "properties": {
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ZipCode": {
                "properties": {
                    "zip_code": {
                        "type": "string"
                    },
                    "locality": {
                        "type": "string"
                    },
                    "federal_entity": {
                        "properties": {
                            "key": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "name": {
                                "type": "string"
                            },
                            "code": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "settlements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Settlement"
                        }
                    },
                    "municipality": {
                        "properties": {
                            "key": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    }
}